Move message handling code into dedicated class
